Move interfaces to private until they are no longer experimental
