+++ lib/private/notification/notification.php
@@ -21,8 +21,6 @@
 
 namespace OC\Notification;
 
-use OCP\Notification\IAction;
-use OCP\Notification\INotification;
 
 class Notification implements INotification {
 	/** @var string */
