+++ lib/public/iservercontainer.php
@@ -459,7 +459,7 @@ interface IServerContainer {
 	/**
 	 * Get the Notification Manager
 	 *
-	 * @return \OCP\Notification\IManager
+	 * @return \OC\Notification\IManager
 	 * @since 8.2.0
 	 */
 	public function getNotificationManager();
