+++ tests/lib/server.php
@@ -105,7 +105,7 @@ class Server extends \Test\TestCase {
 			['NavigationManager', '\OC\NavigationManager'],
 			['NavigationManager', '\OCP\INavigationManager'],
 			['NotificationManager', '\OC\Notification\Manager'],
-			['NotificationManager', '\OCP\Notification\IManager'],
+			['NotificationManager', '\OC\Notification\IManager'],
 			['UserCache', '\OC\Cache\File'],
 			['UserCache', '\OCP\ICache'],
